Add files via upload
Changed the result page to actually work! Huzzah.

Also added some styling to it and to the table of all players. 

In result took the team_id from the post function and entered into the url so that the url would be generated from it. Then used the foreach to get the data only from picks and input it into the table.