Add files via upload
Made some updates to the team.php page by making the entries only allowed to be number values. Currently, the number value is the total amount of teams in the game, but this will change. I will need this to be modified by using the API. If I can create a call and input it into the DB, then query the DB to see what the current total players is then I can just have a php call to the DB to see what the max num is and block anything higher. Have to add a block on the submit button should the entry above be blank.

Also added some styling to the team.php page, mainly some bootstrap styling to fit into the rest of the web pages.